+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergCatalogHandler.java
@@ -32,6 +32,7 @@ import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import jakarta.enterprise.inject.Instance;
 import java.io.Closeable;
+import java.time.Clock;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.ArrayList;
@@ -72,6 +73,7 @@ import org.apache.iceberg.rest.requests.CreateTableRequest;
 import org.apache.iceberg.rest.requests.CreateViewRequest;
 import org.apache.iceberg.rest.requests.RegisterTableRequest;
 import org.apache.iceberg.rest.requests.RenameTableRequest;
+import org.apache.iceberg.rest.requests.ReportMetricsRequest;
 import org.apache.iceberg.rest.requests.UpdateNamespacePropertiesRequest;
 import org.apache.iceberg.rest.requests.UpdateTableRequest;
 import org.apache.iceberg.rest.responses.ConfigResponse;
@@ -121,6 +123,7 @@ import org.apache.polaris.service.events.EventAttributeMap;
 import org.apache.polaris.service.events.EventAttributes;
 import org.apache.polaris.service.http.IcebergHttpUtil;
 import org.apache.polaris.service.http.IfNoneMatch;
+import org.apache.polaris.service.reporting.PolarisMetricsReporter;
 import org.apache.polaris.service.types.NotificationRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -195,6 +198,10 @@ public abstract class IcebergCatalogHandler extends CatalogHandler implements Au
   protected abstract StorageAccessConfigProvider storageAccessConfigProvider();
 
   protected abstract EventAttributeMap eventAttributeMap();
+
+  protected abstract PolarisMetricsReporter metricsReporter();
+
+  protected abstract Clock clock();
 
   // Catalog instance will be initialized after authorizing resolver successfully resolves
   // the catalog entity.
@@ -650,6 +657,16 @@ public abstract class IcebergCatalogHandler extends CatalogHandler implements Au
     }
     return baseCatalog instanceof SupportsNotifications notificationCatalog
         && notificationCatalog.sendNotification(identifier, request);
+  }
+
+  public void reportMetrics(TableIdentifier identifier, ReportMetricsRequest request) {
+
+    authorizeBasicTableLikeOperationOrThrow(
+        PolarisAuthorizableOperation.REPORT_METRICS,
+        PolarisEntitySubType.ICEBERG_TABLE,
+        identifier);
+
+    metricsReporter().reportMetric(catalogName(), identifier, request.report(), clock().instant());
   }
 
   /**